Great work!
Overall this is a great approach to the assignment! To improve, see how much code you can get rid of that is duplicated and ask yourself, is each of these functions necessary? Why or why not? Sometimes in development the thing that will set you apart is your efforts to minimize your code and get more done with less. Great work, man!